Export firebaseExperimentDescription
