+++ docs-devsite/remote-config.firebaseexperimentdescription.md
@@ -1,0 +1,78 @@
+Project: /docs/reference/js/_project.yaml
+Book: /docs/reference/_book.yaml
+page_type: reference
+
+{% comment %}
+DO NOT EDIT THIS FILE!
+This is generated by the JS SDK team, and any local changes will be
+overwritten. Changes should be made in the source code at
+https://github.com/firebase/firebase-js-sdk
+{% endcomment %}
+
+# FirebaseExperimentDescription interface
+Defines experiment and variant attached to a config parameter.
+
+<b>Signature:</b>
+
+```typescript
+export interface FirebaseExperimentDescription 
+```
+
+## Properties
+
+|  Property | Type | Description |
+|  --- | --- | --- |
+|  [affectedParameterKeys](./remote-config.firebaseexperimentdescription.md#firebaseexperimentdescriptionaffectedparameterkeys) | string\[\] |  |
+|  [experimentId](./remote-config.firebaseexperimentdescription.md#firebaseexperimentdescriptionexperimentid) | string |  |
+|  [experimentStartTime](./remote-config.firebaseexperimentdescription.md#firebaseexperimentdescriptionexperimentstarttime) | string |  |
+|  [timeToLiveMillis](./remote-config.firebaseexperimentdescription.md#firebaseexperimentdescriptiontimetolivemillis) | string |  |
+|  [triggerTimeoutMillis](./remote-config.firebaseexperimentdescription.md#firebaseexperimentdescriptiontriggertimeoutmillis) | string |  |
+|  [variantId](./remote-config.firebaseexperimentdescription.md#firebaseexperimentdescriptionvariantid) | string |  |
+
+## FirebaseExperimentDescription.affectedParameterKeys
+
+<b>Signature:</b>
+
+```typescript
+affectedParameterKeys?: string[];
+```
+
+## FirebaseExperimentDescription.experimentId
+
+<b>Signature:</b>
+
+```typescript
+experimentId: string;
+```
+
+## FirebaseExperimentDescription.experimentStartTime
+
+<b>Signature:</b>
+
+```typescript
+experimentStartTime: string;
+```
+
+## FirebaseExperimentDescription.timeToLiveMillis
+
+<b>Signature:</b>
+
+```typescript
+timeToLiveMillis: string;
+```
+
+## FirebaseExperimentDescription.triggerTimeoutMillis
+
+<b>Signature:</b>
+
+```typescript
+triggerTimeoutMillis: string;
+```
+
+## FirebaseExperimentDescription.variantId
+
+<b>Signature:</b>
+
+```typescript
+variantId: string;
+```
