+++ packages/remote-config/src/client/rest_client.ts
@@ -18,7 +18,8 @@
 import {
   CustomSignals,
   FetchResponse,
-  FirebaseRemoteConfigObject
+  FirebaseRemoteConfigObject,
+  FirebaseExperimentDescription
 } from '../public_types';
 import {
   RemoteConfigFetchClient,
@@ -46,32 +47,6 @@ interface FetchRequestBody {
   language_code: string;
   custom_signals?: CustomSignals;
   /* eslint-enable camelcase */
-}
-
-/**
- * Defines experiment and variant attached to a config parameter.
- */
-export interface FirebaseExperimentDescription {
-  // A string of max length 22 characters and of format: _exp_<experiment_id>
-  experimentId: string;
-
-  // The variant of the experiment assigned to the app instance.
-  variantId: string;
-
-  // When the experiment was started.
-  experimentStartTime: string;
-
-  // How long the experiment can remain in STANDBY state. Valid range from 1 ms
-  // to 6 months.
-  triggerTimeoutMillis: string;
-
-  // How long the experiment can remain in ON state. Valid range from 1 ms to 6
-  // months.
-  timeToLiveMillis: string;
-
-  // A repeated of Remote Config parameter keys that this experiment is
-  // affecting the value of.
-  affectedParameterKeys?: string[];
 }
 
 /**
